Tag policies

+++ app/Http/Controllers/TagController.php
@@ -26,6 +26,9 @@ class TagController extends Controller
      */
     public function create()
     {
+        // authorize user
+        $this->authorize('create', Tag::class);
+
         // create a tag
         return view('tags.create');
     }
@@ -35,6 +38,9 @@ class TagController extends Controller
      */
     public function store(StoreTagRequest $request)
     {
+        // authorize user
+        $this->authorize('create', Tag::class);
+
         // store a tag
         $tag = Tag::create($request->validated());
 
@@ -55,6 +61,9 @@ class TagController extends Controller
      */
     public function edit(Tag $tag)
     {
+        // authorize user
+        $this->authorize('update', Tag::class);
+
         // edit a tag
         $games = Game::all();
         return view('tags.edit', compact('tag', 'games'));
@@ -65,6 +74,9 @@ class TagController extends Controller
      */
     public function update(UpdateTagRequest $request, Tag $tag)
     {
+        // authorize user
+        $this->authorize('update', Tag::class);
+
         // update a tag
         $tag->update($request->validated());
 
@@ -76,6 +88,9 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        // authorize user
+        $this->authorize('delete', Tag::class);
+
         // delete a tag
         $tag->delete();
